Remove myself.
While at the time the letter was published, I supported the call for the FSF
board to reconsider his reappointment to the board, they have considered it
a final decision, and I do not think it is useful or helpful to free software
or the continuity of the FSF to continue this particular fight.

Additionally, the FSF board have allocated a seat to be chosen by the FSF
employees to represent their interests, which was one of my primary governance
concerns with the FSF.  That concession has been quite convincing to me that
the goal is to have a much more balanced FSF board.

While I agree with many of the allegations presented in the letter which
present a thesis as to why RMS may be unfavorable to have on the FSF board,
I have also concluded that it is appropriate, and frankly vital, to allow him
to be involved in the planning for the future of the embodiment of his life's
work, regardless of his unfortunate past behaviors.

Given the above, I choose to remove my signature, as it calls for continued
removal of RMS, while I consider this a settled matter.  I do not, however,
agree with or endorse many instances of his past behavior, and will continue
to be critical of future behavioral issues as I deem appropriate.